+++ src/components/auth/page/index.ts
@@ -1,1 +1,1 @@
-export { default as Login } from './Login';
+export { default as LoginPage } from './LoginPage';
