+++ src/components/home/page/index.ts
@@ -1,1 +1,1 @@
-export { default as Home } from './Home';
+export { default as HomePage } from './HomePage';
