+++ src/navigation/RootNavigation.tsx
@@ -2,7 +2,7 @@ import { NavigationContainer, useNavigation } from '@react-navigation/native';
 import { createStackNavigator } from '@react-navigation/stack';
 import type { StackNavigationProp } from '@react-navigation/stack';
 
-import { Login } from '@/components/auth/page';
+import { LoginPage } from '@/components/auth/page';
 import { LOGIN, MAIN_BOTTOM_TAB, SEARCH_NAVIGATION } from '@/constants';
 import { MainBottomTabNavigation, SearchNavigation } from '@/navigation';
 import { useAppSelect } from '@/store';
@@ -20,7 +20,7 @@ const RootNavigation = () => {
   return (
     <NavigationContainer>
       <Stack.Navigator screenOptions={screenOption}>
-        {!accessToken && <Stack.Screen name={LOGIN} component={Login} />}
+        {!accessToken && <Stack.Screen name={LOGIN} component={LoginPage} />}
         <Stack.Screen name={MAIN_BOTTOM_TAB} component={MainBottomTabNavigation} />
         <Stack.Screen name={SEARCH_NAVIGATION} component={SearchNavigation} />
       </Stack.Navigator>
