+++ src/navigation/SearchNavigation.tsx
@@ -1,6 +1,6 @@
 import { createStackNavigator } from '@react-navigation/stack';
 
-import { SubwaySearch } from '@/components/search/page';
+import { SubwaySearchPage } from '@/components/search/page';
 import { SUBWAY_SEARCH } from '@/constants';
 
 const Stack = createStackNavigator();
@@ -12,7 +12,7 @@ const screenOption = {
 const SearchNavigation = () => {
   return (
     <Stack.Navigator screenOptions={screenOption}>
-      <Stack.Screen name={SUBWAY_SEARCH} component={SubwaySearch} />
+      <Stack.Screen name={SUBWAY_SEARCH} component={SubwaySearchPage} />
     </Stack.Navigator>
   );
 };
